pyproject.toml
[project]
name = "person-rlf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bitsandbytes>=0.43.3",
    "easydict>=1.13",
    "fire>=0.6.0",
    "ftfy>=6.2.3",
    "hydra-core>=1.3.2",
    "iprogress>=0.4",
    "ipykernel>=6.29.5",
    "ipython>=8.27.0",
    "ipywidgets>=8.1.5",
    "lightning>=2.4.0",
    "matplotlib>=3.9.2",
    "nltk>=3.9.1",
    "nvitop>=1.3.2",
    "omegaconf>=2.3.0",
    "peft>=0.13.2",
    "pillow>=10.4.0",
    "prettytable>=3.11.0",
    "ruff>=0.6.5",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "tensorboard>=2.17.1",
    "torch-tb-profiler>=0.4.3",
    "torchmetrics>=1.4.2",
    "torchvision>=0.19.1",
    "transformers>=4.44.2",
    "wandb>=0.18.5",
]
